Fix incorrect FIND usage for NUSERS, see below.
FIND /NUSERS/,1- will find multiple lines that include the string NUSERS:
the EQU itself, and then a couple of divisions further down. On completion,
the current location is set to the *last* match, not the first. So I end
up with the wrong number of terminals (the default 4), and an incorrect
buffer allocation further down.