Fix ASSIGN error alluded to in previous commit.
When SEARCH.CSS attempted to search a file that had a nonzero protection
key, e.g. ff00 (read only), it would throw up an ASSIGN error. Solution
is to assign with with access mode SRO (shared read only). That won't
stop it blowing up on files that are keyed ffff or 00ff, but at least it
should handle the most common problem.

I tried putting the search command in an inline $JOB/$TERM construct,
but that doesn't work because the whole thing is already in a job, and
jobs cannot be nested.