Move MTM port determination into HandleSimH_Startup.
That leaves the choice between os32.ini and supnik.ini entirely in the
hands of OS32-FTPd. More importantly, the same ini file is used for the
DO command as for the purposes of finding out the MTM port. Previously,
it would always read os32.ini, even when later switching to supnik.ini.

Consequently, inifile can be taken out of the ::OS32::config namespace,
and removed from ftpd.config.