PR #2 - Increase MTM NUSERS to 10, and add two batch streams.
Also increase NCSBFS to 15 as recommended in the MTM System Planning and
Operator Reference. Otherwise stuff locks up at FIRSTIN, and any time
you try to compile stuff.

Interestingly, this seems to reduce EOU deadlocking, because now I can
EOU in two MTM sessions. But it still locks up if a third session tries
to join in. I'll leave the safety-checks in place.